databricks/requirements.txt
# MUST NOT install databricks
databricks-cli~=0.18.0
databricks-sdk~=0.40.0
python-dotenv~=1.0.1

azure-communication-email==1.0.0

tabulate~=0.9.0
requests~=2.32.3
peewee~=3.18.3

